[#1424] extracted custom MapView and moved it to offline package
