+++ offlinemaps/src/main/java/org/akvo/flow/offlinemaps/di/OfflineFeatureComponent.java
@@ -19,6 +19,7 @@
 
 package org.akvo.flow.offlinemaps.di;
 
+import org.akvo.flow.offlinemaps.presentation.FlowMapViewImpl;
 import org.akvo.flow.offlinemaps.presentation.dialog.OfflineMapsDialog;
 import org.akvo.flow.offlinemaps.presentation.download.OfflineMapDownloadActivity;
 import org.akvo.flow.offlinemaps.presentation.list.OfflineAreasListActivity;
@@ -37,4 +38,6 @@ public interface OfflineFeatureComponent {
     void inject(OfflineMapDownloadActivity activity);
 
     void inject(OfflineMapsDialog dialog);
+
+    void inject(FlowMapViewImpl flowMapView);
 }
