+++ offlinemaps/src/main/java/org/akvo/flow/offlinemaps/presentation/FlowMapView.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2019 Stichting Akvo (Akvo Foundation)
+ * Copyright (C) 2019 Stichting Akvo (Akvo Foundation)
  *
  * This file is part of Akvo Flow.
  *
@@ -15,38 +15,15 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
- *
  */
 
-package org.akvo.flow.presentation.datapoints.map;
+package org.akvo.flow.offlinemaps.presentation;
 
-import org.akvo.flow.domain.entity.DataPoint;
+import org.akvo.flow.offlinemaps.domain.entity.MapInfo;
 
-import java.util.List;
+public interface FlowMapView {
 
-interface DataPointsMapView {
+    void displayOfflineMap(MapInfo mapInfo);
 
-    void showProgress();
-
-    void hideProgress();
-
-    void displayDataPoints(List<DataPoint> dataPoints);
-
-    void showSyncedResults(int numberOfSyncedItems);
-
-    void showErrorAssignmentMissing();
-
-    void showErrorNoNetwork();
-
-    void showErrorSync();
-
-    void showNoDataPointsToSync();
-
-    void hideMenu();
-
-    void showMonitoredMenu();
-
-    void showNonMonitoredMenu();
-
-    void showFab();
+    void displayUserLocation();
 }
