+++ offlinemaps/src/main/res/drawable-hdpi/marker.png
