+++ offlinemaps/src/main/res/drawable-mdpi/marker.png
