+++ offlinemaps/src/main/res/drawable-xhdpi/marker.png
