+++ offlinemaps/src/main/res/drawable-xxhdpi/marker.png
