+++ offlinemaps/src/main/res/drawable-xxxhdpi/marker.png
