+++ offlinemaps/src/main/res/layout/symbol_layer_info_window_layout_callout.xml
